Get rid of asynchronicity
Turns out there is no need to poll server with asynchronous messages to determine when the server is listening. Provided that we've called connect with JuliaClient, the client will wait for the server to start before sending its message. If we call socket.recv directly after sending, we will have the correct pairing of the response with the message that we want.